Optimize db queries
