+++ workers/loc.api/sync/dao/dao.js
@@ -90,17 +90,7 @@ class DAO {
   /**
    * @abstract
    */
-  async checkAuthInDb () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
   async findInCollBy () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
-  async getActiveUsers () { throw new ImplementationError() }
 
   /**
    * @abstract
@@ -125,26 +115,6 @@ class DAO {
   /**
    * @abstract
    */
-  async getSubUsersByMasterUser () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
-  async insertOrUpdateUser () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
-  async updateUserByAuth () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
-  async deactivateUser () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
   async getElemsInCollBy () { throw new ImplementationError() }
 
   /**
@@ -161,11 +131,6 @@ class DAO {
    * @abstract
    */
   async updateStateOf () { throw new ImplementationError() }
-
-  /**
-   * @abstract
-   */
-  async getFirstElemInCollBy () { throw new ImplementationError() }
 
   /**
    * @abstract
