+++ workers/loc.api/sync/dao/helpers/utils.js
@@ -1,13 +1,13 @@
 'use strict'
 
-const { pick } = require('lodash')
 const {
   AuthError
 } = require('bfx-report/workers/loc.api/errors')
 
 const {
-  SubAccountCreatingError
+  ObjectMappingError
 } = require('../../../errors')
+
 const { deserializeVal } = require('./serialization')
 
 const mixUserIdToArrData = (
@@ -60,28 +60,37 @@ const convertDataType = (
   return arr
 }
 
-const pickUserData = (user) => {
-  return {
-    ...pick(
-      user,
-      [
-        'apiKey',
-        'apiSecret',
-        'email',
-        'timezone',
-        'username',
-        'id'
-      ]
-    )
-  }
-}
+const mapObjBySchema = (obj, schema = {}) => {
+  const err = new ObjectMappingError()
 
-const checkUserId = (user = {}) => {
-  const { _id } = { ...user }
-
-  if (!Number.isInteger(_id)) {
-    throw new SubAccountCreatingError()
+  if (
+    !obj ||
+    typeof obj !== 'object' ||
+    !schema ||
+    typeof schema !== 'object'
+  ) {
+    throw err
   }
+
+  const map = Array.isArray(schema)
+    ? schema.map(item => [item, null])
+    : Object.entries(schema)
+
+  return map.reduce((accum, [key, val]) => {
+    const _val = val && typeof val === 'string' ? val : key
+
+    if (
+      !key ||
+      typeof key !== 'string' ||
+      typeof obj[_val] === 'undefined'
+    ) {
+      throw err
+    }
+
+    accum[key] = obj[_val]
+
+    return accum
+  }, {})
 }
 
 const isContainedSameMts = (
@@ -108,7 +117,6 @@ const isContainedSameMts = (
 module.exports = {
   mixUserIdToArrData,
   convertDataType,
-  pickUserData,
-  checkUserId,
-  isContainedSameMts
+  isContainedSameMts,
+  mapObjBySchema
 }
